# Conflicts:
#	bible/src/main/java/com/library/bible/reservation/model/Reservation.java

+++ bible/src/main/java/com/library/bible/reservation/repository/IReservationRepository.java
@@ -14,8 +14,8 @@ public interface IReservationRepository {
 	List<Reservation> selectAllReserv();
 	
 	//Reservation CRUD
-	Reservation selectReserv(int reservId);
+	Reservation selectReserv(long reservId);
 	void insertReserv(Reservation reservation);
 	void updateReserv(Reservation reservation);
-	int deleteReserv(int reservId);
+	int deleteReserv(long reservId);
 }
